Remove AI driver for infiSTAR update
This removes the AI driver for compatibility with infiSTAR version 419.

The only difference is the plane will crash instead of flying straight
after the player ejects.

InfiSTAR 419 also kicks players for halo jumping when his new keybinds
check is enabled. Make sure to set it to false in AHconfig.sqf:
<pre><code>/*  Check Keybinds ?      */ _CKD =  false; /* true or false
*/ /* checks for additional keybinds.. */</code></pre>
That is until he fixes it so there is an exception for the keybinds
BIS_fnc_halo adds.